// src/lib/api-types.ts
export interface components {
    schemas: {
// ...
        FeedItem: {
            id: string;
            title: string;
            body: Record<string, never> | null;
            /** Format: date-time */
            createdAt: string;
            author: components["schemas"]["FeedUser"];
            soma: components["schemas"]["FeedSoma"];
            media: components["schemas"]["FeedMediaCollection"] | null;
            voteCount: number;
            awardCount: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}